Before:

```
❯ hyperfine -w 1 -r 5 './pants --no-pantsd lint --only=docformatter ::'
Benchmark 1: ./pants --no-pantsd lint --only=docformatter ::
  Time (mean ± σ):      6.907 s ±  0.651 s    [User: 5.204 s, System: 0.876 s]
  Range (min … max):    6.299 s …  7.950 s    5 runs
```

After:

```
❯ hyperfine -w 1 -r 5 './pants --no-pantsd lint --only=docformatter ::'
Benchmark 1: ./pants --no-pantsd lint --only=docformatter ::
  Time (mean ± σ):      6.478 s ±  0.111 s    [User: 4.995 s, System: 0.779 s]
  Range (min … max):    6.369 s …  6.666 s    5 runs
```

[ci skip-rust]
[ci skip-build-wheels]